update crownholder in whoknows
